Adaptive Mesh Implemented ------------------------------------
The adaptive mesh is here and works very nicely. Since the compatibilty with previous versions has been broken, this was also the perfect moment to comment out some obsolete variables.

Now the optimization cycle has started over. updateH() is now so fast that it is not the bottleneck anymore. updateQ() and updateCharges() are the next in line to be optimized.